+++ src/main/java/com/ilich/sb/e_commerce/service/impl/ProductServiceImpl.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.ilich.sb.e_commerce.model.Category;
 import com.ilich.sb.e_commerce.model.Product;
 import com.ilich.sb.e_commerce.repository.IProductRepository;
 import com.ilich.sb.e_commerce.service.IProductService;
@@ -32,8 +33,18 @@ public class ProductServiceImpl implements IProductService {
     }
 
     @Override
-    public void delete(long id) {
-        iProductRepository.deleteById(id);
+    public Product update(long id, Product product) {
+        Product pro = iProductRepository.findById(id).orElseThrow(()-> new RuntimeException("No existe id: "+id));
+        return iProductRepository.save(product);
+    }
+
+    @Override
+    public boolean delete(long id) {
+        if (iProductRepository.existsById(id)) {
+            iProductRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
